Add files via upload
This is the project that helps us to manage the groceries list in our home. You can add new groceries with item name, quantity and brand name . You can edit  and remove them. I had made this project in eclipse with the help of Apache-tomcat server. You can store and retrieve grocery list in database which is done with the help of  jdbc.